APIMF-3324: Error handler implemented
